Move testUtils folder
