+++ sts-mock/src/functions/sts-mock/token/tests/tokenHandler.test.ts
@@ -1,11 +1,11 @@
 import { Logger } from "../../services/logging/logger";
 import { MockLoggingAdapter } from "../../services/logging/tests/mockLoggingAdapter";
-import { buildLambdaContext } from "../../../../testUtils/mockContext";
+import { buildLambdaContext } from "../../../testUtils/mockContext";
 import { MessageName, registeredLogs } from "../registeredLogs";
 import { lambdaHandlerConstructor } from "../tokenHandler";
 import { APIGatewayProxyEvent } from "aws-lambda";
 import { TokenDependencies } from "../handlerDependencies";
-import { buildTokenRequest } from "../../../../testUtils/mockRequest";
+import { buildTokenRequest } from "../../../testUtils/mockRequest";
 import {
   MockTokenSignerErrorResult,
   MockTokenSignerSuccessResult,
